requirements.txt
PyYAML
dotmap
bravado
oauthlib
jinja2
celery
django
